oooooh this stinks!
Areas of Interest.pickle added to data, template code adjusted accordingly.

SmallLandsat file for singular satellite images. no way to use it with collections yet.

Template_gee can now successfully mark an area of interest, save it with a name, download it to drive, analyze it through a numpy arrray.
Only downside here is that it can only download high res to drive and not directly to local.